117.1 Putting it All Together(Section 11: Let's Test Your React Mastery! )
In the last section we put together some code to make a request to the YouTube API.

We're now ready to make use of that inside of our apps.

Yes file.

Remember any time that someone hits enter or submits our search form we want to reach out to YouTube

and try to find some results from their API.

So the first thing we have to do is add a callback method to our class app that will be called any time

that someone submits the search bar form.

So inside of my class up I'll say about on search or say on term submits and I'll make sure I defined

this as an arrow function like so and I'm going to assume that the thing is going to be called with

some term search string.

Now the reason I'm calling this on term Samit is that inside of our search bar remember we had called

the state inside of your term term being short for search term.

So I think that kind of makes sense just to carry that terminology through our application.

So then on my search bar can phone right here I'm going to add this callback as a prop will say how

about like on form submit.

I want to call this start on term submit

remember.

And we add a prop or kind of a callback name to a component that you and I create we can use any prop

name that you and I want to use.

So we could have each just as easily called out on forms submit could have been Call me when it submitted

know anything we want it to be.

But I think that on unformed Smit makes a pretty good amount of sense.

Does he know if I was making this application for myself I probably would have matched this to the name

of the actual callback.

So I probably would have called it something like On term Samit and made the two identical.

The only reason I'm specifically showing a different product name right here is just so you recall that

this and this do not have to be identical.

That's the only reason I'm not making the same otherwise if this was my own project that was making

for my own purpose.

I would probably give them the same exact name.

All right.

So now back inside of my search bar I'm going to find my on for summit number.

We've got to do right here.

We want to make sure that we call that prop any time that these are Smyth's the form.

So inside of your all say this stock props start on form submit

and I'm going to call that with this state term so that our parent component is told what the current

search term is.

So now whenever we submit the form this function right or should be cold with the appropriate search

term Just make sure that's the case on that.

On a quick console log in console log out the term.

All right so let's flip back over.

I'll enter in some search term hit enter and I see the appropriate counsel lock.

Perfect.