114. Axios Version Bug &amp; 400 "Required Parameter: part" Error 117.2 Putting it All Together(Section 11: Let's Test Your React Mastery! )
In the upcoming lecture, we will be creating an axios config instance to call the YouTube API. Starting with release v0.19.0 there is a bug that fails to merge the parameters from the config instance. This will cause a 400 error failure with the message "Required Parameter : part"

Here are two ways we can deal with this:

1. Install a specific version of axios:

npm install axios@0.18.1

2. Move the Params object to the App.js axios call

117.2 Putting it All Together

So I think that we're now ready to go ahead and remove that console log import the YouTube API and use

that to actually make a request to YouTube.

So at the top I'm going to add a import statement for simply YouTube and I'll import that from up one

directory so notice the two dots right there.

That means go up into the Cercy folder then go into the API folder and find the file.

YouTube.

So then inside of your I can call the U2 function.

Remember this is now a preconfigured instance of ASIO's.

So I'll say YouTube does get a pass in the route or the path that I want to make a get request to.

In this case I want to access the search and point and then now that we know what the search term is

we can pass in the actual Paramo of Q to tell the YouTube API exactly what search term we want to use.

So as the second argument all out in a parampara object that's going to have a property called Q.

Remember this kid right here is called Q Specifically because that is what that YouTube API documentation

told us to call it simply.

Q Not Kohary 90 like that just Q by itself.

And then our search string will be term like so which is the argument that had been passed into the

on terms Samit function.

Now this right here is a asynchronous API request.

So in order to interact with it we have to use a promise or that async a wait syntax that we saw a little

bit ago but before we worry about that too much let's just try to make a request with what you see right

here and just ensure that we actually see our request come up in the network request log and that we

can actually reach out to YouTube successfully.

So I going to save this file and my search for a file and then I'll flip back over to my application.

I'm going to open up my network request log.

Network not elements and then I'm going to filter by a char and fetch requests.

And then just make this a little bit easier to read.

All click on clear to throw out that requests log.

So now we can go ahead and do a search.

Let's go ahead and do buildings.

When I hit enter.

You'll notice that a request right here pops up.

I can hover over it and I'll see the entire query string on there.

So you see I can't quite move my mouse to show you exactly but you can see at the end of that Eurail

on the first line it says v3 search and then has the little question mark that indicates that we're

about to add on a query string.

So you can see part snippet Max results Phife our API key and then q equals buildings.

So all those different query parameters got added onto the request.

So now we can click on this go to the preview tab right here and it will show you the response we got

from YouTube.

All right.

So inside of here is a item's property.

There are five items zero through four.

And if you click on any of those items and then expand this snippet property you'll see that it has

some information about a very particular YouTube video.

So in this case this video has a title of 17 best building demolitions ever.

Boy that sounds like a amazing video.

I'm going to watch that after this.

And we've got some thumbnails on here that contains some links to images that we can show on the screen.

Now we've also got a video I.D. as well.

You'll notice back up here outside the snippet.

Is this I.D. property inside of it is a video I.D. this video I.D. is how we're going to eventually

get the video to play inside of our application.

So that's going to eventually be pretty relevant for what we're trying to do.

All right so that's it.

We've now successfully made a request to the YouTube API.

So let's take a pause here.

When we come back the next section we're going to start to focus on taking that list of videos and rendering

them as a list inside of our application.