122.1 Styling a List(Section 11: Let's Test Your React Mastery! )
In the last section we got each video item to show what the title of the video it's representing an

image next to it as well.

Now this does not look too good as it stands right now.

So in this video we're going to work on just a little bit of styling to get started.

I'm going to pull open my semantic you eye docs.

Remember you can get you're out semantic dash you white dot com.

And then on the left hand menu under the elements section I'm going to find the link to list inside

of the list documentation or then scroll down a little bit to this kind of a little get hub looking

examplar right here.

It We'll take a look at the source for it.

So the wrapping div..

The div that contains each of these individual items needs to have a class of UI relaxed divided list

and then each individual item has to have a class of item.

So to translate this into our current component hierarchy we're essentially saying that our video list

component has to have a div of UI relax divided lists and then each individual video item has to return

an element that has a top level class name of item.

So let's do that refactor right now and hopefully ill will fix up a little bit of styling.

All right so I'll start at my video list right here.

So here's the div that wraps every individual video item that needs to have a class name of UI relaxed

divided list so relaxed invited list.

And now I'll flip back over to my video item component and now the root div inside of here needs to

have a class name of item.

So I'll put on the video items div a class name of item art.

So there's a couple of more class names we have to put in here.

But let's at least test this as it stands and make sure that we see at least some difference inside

of our application.

So I'll do another search.

And I at least see the little kind of dividing line right there.

That tells me that we're at least on the right path but we still need to do a little bit more than this.

So if I look at the documentation over here again on notice that I have a div inside there the first

settlment is essentially going to be some image or icon that is going to be displayed to the left to

on side.

There's actually another example inside of this file where this page right here that shows us how to

make use of an image on the left hand side.

So let's scroll down a little bit and find that it's all go down to image right here.

So a list item can contain an image.

If I expand this list I'll see that the div with class of item can have an image with a class of UI

avatar image.

Now in this case the class name of Avatar that you see on there is to just get the image to appear as

a circle.

In our case we probably don't want to show these thumbnails as a circle so we are not going to use the

Avatar class name.

Now you'll also notice that the content inside of this list has a div wrapped around it with a class

of content and then inside there is an element with a class of Hetter.

So it's essentially implement all these class names into our video item and I think it's going to help

our styling just a little bit.

All right tobacco-free here.

I'm going to first find my image tag.

I'll add on a class name of UI image and remember I do not need Avatar because all that's going to do

is turn the image into a circle.

And now after that I'm going to add a div I'm going to put my video snippet title inside there and I

will give that div a class name of content I'm going to wrap the title Inside of a another div inside

of your

and the inner div will get a class name of header like so.

So now I've got a very similar structure here of class names to what you can see in this example.

I've got the top level item the image a div with class content and then inside there in element with

class name header.

Now you will notice that they make use of an anchor tag.

In this case we don't have anything clickable per say just yet.

So we're going to render it simply as a div and even though we're using a different element the class

name is what really provides the styling with semantic UI.

So we can switch out the actual element type that we're using and just apply the same class name and

it will still do the appropriate styling to it.

OK.

So here's my video item.

I'm going to save this or then flip back over and we'll do a quick test so I'll search for buildings.

All right so that definitely definitely looks better than what we had before.

On the right hand side I've got my image I have the kind of nicely formatted because it's nice and bold

header on the right hand side.

But now the only issue is that the image is showing up to large and the text right here.

I feel like would be a lot better if it was kind of centered next to the image.

Now unfortunately to do those two changes to limit the size of the image and to center the text right

here we can't use semantic UI for that.