122.2 Styling a List(Section 11: Let's Test Your React Mastery! )
We're going to have to write just a little bit of custom CSSA rather than taking a pause So let's just

get it done right now so we can have one video that contains all the styling stuff we have to do.

All right.

So we're going to write out a little bit of custom styling here.

We could either add a custom style tag to the elements you display in here or we could create a separate

CSSA file.

In this case I'm going to decide to create a separate CSSA file so that we don't clutter up this component

with a bunch of styling rules.

So inside my component's directory I'm going to make a new file called Video item C Ss

and then I'm going to flip back over to the video item and I will import that CSSA file at the very

top.

So I will say simply import slash video item that SS like so now I remember a rule that I like to use

to find the root element that is returned from some component and give it a class name equal to the

component name.

And that just make sure that when I write some CSSA I can style just the GSX inside of your with some

reasonable certainty rather than having some CSSA that it's going to accidentally be applied to all

of the different elements inside my application.

And there's definitely some rules on how effective this is especially when my component is nesting some

other component but it's at least an approach or at least a start.

So on my route div right here I'll add on a class name of video dash item like so it's now over inside

of a video item that says I can target video dash item.

All right.

So inside of here we're going to put in just a little bit of custom CSSA to make sure that the title

right here appears directly centered next to the image and to also make sure that the image is not too

large.

Now we're not going to focus too much on the actual styling that we're doing here because as I've said

many times our focus in this course is react not quite so much the CSA side of things.

So I can't put on a display of flex and I'm going to mark this thing as important.

In general using an important rule is not super heavily advised but this is one case where again not

super heavily worried about the CSSA here.

I just want to get something that looks reasonable.

So then after that I'll do a line items center also with important.

And then I'll do a cursor of pointer and you'll see why we're putting that on.

And just a second.

Then after that I'll do a video dash item and I'm going to select the image that we are rendering inside

of that component.

And I'm going to give it a max width of one hundred and eighty pixels.

So let's just get to make sure that the image doesn't appear too large.

All right so if I now flip back over it appears that the text is correctly centered on the screen.

And if I zoom out 200 percent I don't think that the image style was picked up correctly we might have

to work on the selector we're using for that rule right there.

Yeah I don't see any.

Oh you know what do I put.

Now I got video dash item.

You only have to put on here is video dash item item like so let's try that out and I'll do a save.

OK.

That's much better.

So the reason that was not initially working was that there was probably some other style.

OK here it is right here.

There's some other rule there was setting a max with property with a higher importance than the one

that we were trying to assign.

So by adding on a more specific selector right.

Of item.

Our rule was made more specific than the default one.

And so our silent kicked in for the max width of the image.

OK.

So in total a little bit of styling craziness here but I think that we've now got a list that looks

pretty darn good.

So let's take another quick pause here and we'll continue in the next section.